+++ src/main/java/com/algoanalyzer/domain/problem/api/dto/SolvedAcProblemResponse.java
@@ -13,7 +13,7 @@ import lombok.Setter;
 @Setter
 public class SolvedAcProblemResponse {
     @JsonProperty("problemId")
-    private int problemId;
+    private Long problemId;
     
     @JsonProperty("titleKo")
     private String titleKo;
